remove Old*Kernel types and factories
